Heredoc Reeeeaady inchallah

+++ MasterMerged/parsing/identificators/identity_tools/identool4.c
@@ -6,7 +6,7 @@
 /*   By: ayel-bou <ayel-bou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/04 03:19:33 by ayel-bou          #+#    #+#             */
-/*   Updated: 2025/08/13 17:41:26 by ayel-bou         ###   ########.fr       */
+/*   Updated: 2025/08/14 21:21:48 by ayel-bou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,9 +79,12 @@ static int	open_heredoc(t_token *id_class, t_token *curr, t_data *data)
 	del = scrap_del(get_delimiter(curr));
 	if (!del)
 		return (free(gename), 0);
-	data->here_fd = open(gename, O_CREAT | O_RDWR, 0777);
+	data->here_fd = open(gename, O_CREAT | O_WRONLY, 0777);
 	if (data->here_fd == -1)
 		return (free(del), free(gename), 0);
+	data->here_read_fd = open(gename, O_RDONLY, 0777);
+	if (data->here_read_fd == -1)
+		return (close(data->here_fd), free(del), free(gename), 0);
 	unlink(gename);
 	if (!here_doc_ops(id_class, data, del))
 		return (free(del), free(gename), 0);
